lib/ukswrand: split ukswrand implementation into ukswrand and ukhwrand
