+++ lib/ukrand/Makefile.uk
@@ -1,0 +1,17 @@
+$(eval $(call addlib_s,libukrand,$(CONFIG_LIBUKRAND)))
+
+CINCLUDES-$(CONFIG_LIBUKRAND)	+= -I$(LIBUKRAND_BASE)/include
+CXXINCLUDES-$(CONFIG_LIBUKRAND) += -I$(LIBUKRAND_BASE)/include
+
+LIBUKRAND_SRCS-$(CONFIG_LIBUKRAND_MWC) += $(LIBUKRAND_BASE)/mwc.c
+LIBUKRAND_SRCS-$(CONFIG_LIBUKRAND_CHACHA) += $(LIBUKRAND_BASE)/chacha.c
+LIBUKRAND_SRCS-$(CONFIG_LIBUKRAND_DEVFS) += $(LIBUKRAND_BASE)/dev.c
+LIBUKRAND_SRCS-y += $(LIBUKRAND_BASE)/swrand.c
+LIBUKRAND_SRCS-y += $(LIBUKRAND_BASE)/getrandom.c
+LIBUKRAND_SRCS-y += $(LIBUKRAND_BASE)/rand.c
+LIBUKRAND_SRCS-$(CONFIG_LIBUKRAND_TRUE_RANDOMNESS) += $(LIBUKRAND_BASE)/hwrand.c
+
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBUKRAND) += getrandom-3
+
+
+
