+++ lib/ukrand/exportsyms.uk
@@ -5,4 +5,5 @@ uk_swrand_def
 uk_swrand_init_r
 uk_swrand_randr_r
 uk_swrandr_gen_seed32
-uk_swrand_fill_buffer
+uk_swrand_generate_bytes
+uk_hwrand_generate_bytes
