+++ lib/ukrand/include/sys/random.h
