+++ lib/ukrand/mwc.c
