+++ lib/ukrand/rand.c
@@ -1,8 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /*
- * Authors: Simon Kuenzer <simon.kuenzer@neclab.eu>
- *
- * Copyright (c) 2018, NEC Europe Ltd., NEC Corporation. All rights reserved.
+ * Copyright (c) 2019, University Politehnica of Bucharest. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -29,52 +27,30 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-
-#ifndef __UK_SWRAND__
-#define __UK_SWRAND__
-
-#include <sys/types.h>
-#include <uk/arch/types.h>
-#include <uk/plat/lcpu.h>
+#include <string.h>
 #include <uk/config.h>
-#include <uk/plat/time.h>
+#include <uk/print.h>
+#include <uk/init.h>
+#include "uk/hwrand.h"
+#include "uk/swrand.h"
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+static int _uk_random_generator_init() {
+	#ifdef CONFIG_LIBUKRAND_TRUE_RANDOMNESS
+		// TODO HW randomness init
+		if (!has_RDRAND()) {
+			uk_pr_err("No support for RDRAND\n");
 
-#define UK_SWRAND_CTOR_PRIO	1
+			#undef CONFIG_LIBUKRAND_TRUE_RANDOMNESS
+			#define CONFIG_LIBUKRAND_PSEUDO_RANDOMNESS
 
-struct uk_swrand;
+			return _uk_swrand_init();
+		}
 
-extern struct uk_swrand uk_swrand_def;
+		return 0;
 
-void uk_swrand_init_r(struct uk_swrand *r, unsigned int seedc,
-			const __u32 seedv[]);
-__u32 uk_swrand_randr_r(struct uk_swrand *r);
+	#endif
 
-__u32 uk_swrandr_gen_seed32(void);
-/* Uses the pre-initialized default generator  */
-/* TODO: Add assertion when we can test if we are in interrupt context */
-/* TODO: Revisit with multi-CPU support */
-static inline __u32 uk_swrand_randr(void)
-{
-	unsigned long iflags;
-	__u32 ret;
-
-	iflags = ukplat_lcpu_save_irqf();
-	ret = uk_swrand_randr_r(&uk_swrand_def);
-	ukplat_lcpu_restore_irqf(iflags);
-
-	return ret;
+	return _uk_swrand_init();
 }
 
-ssize_t uk_swrand_fill_buffer(void *buf, size_t buflen);
-int RDRAND_bytes(unsigned char *buffer, size_t buflen);
-
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* __UK_SWRAND__ */
+uk_early_initcall(_uk_random_generator_init);
