+++ lib/uksp/Config.uk
@@ -36,7 +36,7 @@ config LIBUKSP_VALUE_USECONSTANT
 
 config LIBUKSP_VALUE_RANDOM
 	bool "Random variable"
-	select LIBUKSWRAND
+	select LIBUKRAND
 endchoice
 
 config LIBUKSP_VALUE_CONSTANT
