--- lib/ukswrand/Config.uk
@@ -1,65 +1,0 @@
-menuconfig LIBUKSWRAND
-	bool "ukswrand: Software random number generator"
-	select LIBUKDEBUG
-	default n
-
-if LIBUKSWRAND
-
-choice
-	prompt "Support"
-	default LIBUKSWRAND_PSEUDO_RANDOMNESS
-
-config LIBUKSWRAND_PSEUDO_RANDOMNESS
-	bool "PRNG"
-	help
-		"Pseudo random number generator support"
-
-config LIBUKSWRAND_TRUE_RANDOMNESS
-	depends on ARCH_X86_64
-	bool "TRNG (RDRAND)"
-	help
-		"True random number generator support based on RDRAND instruction"
-endchoice
-
-choice
-	prompt "Algorithm"
-	default LIBUKSWRAND_CHACHA
-
-config LIBUKSWRAND_MWC
-	bool "Multiply-with-carry"
-	help
-		Use multiply-with-carry algorithm
-
-config LIBUKSWRAND_CHACHA
-	bool "ChaCha20"
-	help
-		Use ChaCha20 algorithm
-endchoice
-
-choice
-    prompt "Initial seed"
-    default LIBUKSWRAND_INITIALSEED_TIME
-
-config LIBUKSWRAND_INITIALSEED_TIME
-    bool "Platform timestamp"
-
-config LIBUKSWRAND_INITIALSEED_RDRAND
-    depends on ARCH_X86_64 && (MARCH_X86_64_COREI7AVXI)
-    bool "`rdrand` instruction"
-
-config LIBUKSWRAND_INITIALSEED_USECONSTANT
-    bool "Compiled-in constant"
-endchoice
-
-config LIBUKSWRAND_INITIALSEED_CONSTANT
-    int "Initial seed constant"
-    depends on LIBUKSWRAND_INITIALSEED_USECONSTANT
-    default 23
-
-config LIBUKSWRAND_DEVFS
-	bool "Register random and urandom device to devfs"
-	select LIBDEVFS
-	select LIBVFSCORE
-	default n
-
-endif
